web: fixed bug with exchange rates
it was an int instead of a float for graphql

+++ api/src/schema/payments/currency.ts
@@ -1,5 +1,5 @@
 import { prop as Property, getModelForClass, modelOptions } from '@typegoose/typegoose';
-import { ObjectType, Field, Int } from 'type-graphql';
+import { ObjectType, Field, Float } from 'type-graphql';
 
 @ObjectType({ description: 'currency data' })
 @modelOptions({ schemaOptions: { collection: 'currencies' } })
@@ -8,7 +8,7 @@ export default class Currency {
   @Property({ required: true })
   name: string;
 
-  @Field(_type => Int, { description: 'exchange rate to usd' })
+  @Field(_type => Float, { description: 'exchange rate to usd' })
   @Property({ required: true, type: Number })
   exchangeRate: number;
